some refactorings in color handler
--HG--
extra : convert_revision : svn%3Ad131b684-3811-0410-b0a9-b858155d7784/trunk%40123

+++ src/ColorSpecialHandler.h
@@ -32,7 +32,7 @@ using std::vector;
 
 class ColorSpecialHandler : public SpecialHandler
 {
-	typedef vector<double> RGB;
+	typedef vector<float> RGB;
 
    public:
 		bool process (istream &is);
